+++ routes/searchRoutes.js
@@ -1,0 +1,7 @@
+const express = require('express');
+const router = express.Router();
+const searchController = require('../controllers/searchController');
+
+router.get('/', searchController.getSearchResult);
+
+module.exports = router;
